conductor(checkpoint): Checkpoint end of Phase 1: Refactoring
