+++ android/features/music-player/src/main/java/io/newm/feature/musicplayer/MusicPlayerSlider.kt
@@ -106,6 +106,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.lerp
 import io.newm.core.theme.DarkPink
 import io.newm.core.theme.DarkViolet
+import io.newm.feature.musicplayer.components.MusicPlayerBrush
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.coroutineScope
